Builds/Lunar-Forge/pyproject.toml
[project]
name = "lunar-forge"
version = "0.1.0"
description = "Profile-based build system for Luna Engine"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "textual>=0.40.0",
]

[project.scripts]
lunar-forge = "lunar_forge.build:main"

[tool.setuptools.packages.find]
where = ["."]
